삭제 유효성 검사 및 리렌더링 구현

+++ frontend/src/pages/ItemList.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import itemApi from '../apis/itemApi';
 import ItemCard from '../components/ItemCard';
 import { useNavigate } from 'react-router-dom';
@@ -8,8 +8,13 @@ const ItemList = () => {
 
   const nav = useNavigate();
 
+  const [reload, setReload] = useState(0);
+
   // 체크된 아이디를 저장할 state 변수
   const [checkedItems, setCheckedItems] = useState([]);
+
+  // 전체 체크박스를 저장할 useRef
+  const checkedAll = useRef([]);
 
   // 목록 삭제 여부를 판단할 state 변수
   const [isDelete, setIsDelete] = useState(false);
@@ -23,12 +28,13 @@ const ItemList = () => {
       try {
         const res = await itemApi.getAll();
         setItemList(res);
+        checkedAll.current = res.map(d => d.id);
       } catch (e) {
         handleError(e, false);
       }
     }
     fetchItems();
-  }, []);
+  }, [reload]);
 
   // 개별 체크박스 클릭 시 실행 함수
   const checkItem = e => {
@@ -42,14 +48,32 @@ const ItemList = () => {
     }
   };
 
+  // 전체 체크박스 체크 시 실행 함수
+  const checkAll = e => {
+    setCheckedItems(e.target.checked ? checkedAll.current : []);
+  }
+
   // 게시글 목록 삭제
   const deleteList = async idList => {
-    const res = await itemApi.deleteList(idList);
-    alert(res);
-    setCheckedItems([]);
+    if(!idList || idList.length === 0){
+      alert("삭제할 목록을 선택해주세요.");
+      return;
+    }
+    if(!confirm("삭제하시겠습니까?")){
+      return;
+    }
+    try{
+      const res = await itemApi.deleteList(idList);
+      setReload(prev => prev + 1);
+      alert(res);
+      setCheckedItems([]);
+    } catch(e){
+      handleError(e, true)
+    }
   }
 
   console.log(checkedItems)
+  console.log(checkedAll)
 
   return (
     <div>
@@ -58,6 +82,15 @@ const ItemList = () => {
         
         {isDelete ? (
           <div className="flex items-center gap-2">
+            <label className="flex h-9 items-center gap-1.5 rounded-md bg-blue-600 px-4 text-sm font-medium text-white transition-colors hover:bg-blue-700 cursor-pointer">
+              <input 
+                type="checkbox"
+                className="h-4 w-4 rounded border-white"
+                checked={checkedAll.current.length === checkedItems.length}
+                onChange={e => checkAll(e)}
+              />
+              <span>전체</span>
+            </label>
             <button 
               onClick={() => setIsDelete(false)}
               className="flex h-9 items-center gap-1.5 rounded-md bg-gray-500 px-4 text-sm font-medium text-white transition-colors hover:bg-gray-600 active:bg-gray-400"
